refactor organization file proyect
