--- app/src/main/java/com/example/mpagos/ui/util/domain/UiState.kt
@@ -1,2 +1,0 @@
-package com.example.mpagos.ui.util.domain
-
