+++ app/src/main/java/com/example/mpagos/util/domain/UiState.kt
@@ -1,0 +1,2 @@
+package com.example.mpagos.util.domain
+
